node_modules removed, starrating moved to components/gradeboxLependin
